requirements.txt
pypdf
python-docx
textract==1.6.3
requests
beautifulsoup4
langchain-openai
trustcall
dotenv